fix(input): input h5 demo拆解
