+++ src/packages/input/demo.tsx
@@ -1,10 +1,21 @@
-import React, { useState } from 'react'
-import { Tips, Close, Eye, Marshalling } from '@nutui/icons-react'
-import { Input } from './input'
-import Form from '@/packages/form'
+import React from 'react'
+
 import { useTranslate } from '../../sites/assets/locale'
-import Button from '@/packages/button'
-import Toast from '@/packages/toast'
+
+import Demo1 from './demos/h5/demo1'
+import Demo2 from './demos/h5/demo2'
+import Demo3 from './demos/h5/demo3'
+import Demo4 from './demos/h5/demo4'
+import Demo5 from './demos/h5/demo5'
+import Demo6 from './demos/h5/demo6'
+import Demo7 from './demos/h5/demo7'
+import Demo8 from './demos/h5/demo8'
+import Demo9 from './demos/h5/demo9'
+import Demo10 from './demos/h5/demo10'
+import Demo11 from './demos/h5/demo11'
+import Demo12 from './demos/h5/demo12'
+import Demo13 from './demos/h5/demo13'
+import Demo14 from './demos/h5/demo14'
 
 const InputDemo = () => {
   const [translated] = useTranslate({
@@ -75,134 +86,38 @@ const InputDemo = () => {
       wordCount: 'Word count',
     },
   })
-  const formatter = (value: string) => value.replace(/\d/g, '')
-  const [val, setVal] = useState('NutUI React')
-  const [inputType, setInputType] = useState('password')
-  const [currentLength, setCurrentLength] = useState(0)
 
   return (
     <>
       <div className="demo" style={{ paddingBottom: '20px' }}>
         <h2>{translated.basic}</h2>
-        <Input
-          placeholder={translated.text}
-          onChange={(v) => {
-            console.log('onChange', v)
-          }}
-        />
+        <Demo1 />
         <h2>{translated.uncontrolled}</h2>
-        <Input defaultValue="NutUI React" placeholder={translated.text} />
+        <Demo2 />
         <h2>{translated.controlled}</h2>
-        <Input
-          value={val}
-          onChange={(val) => setVal(val)}
-          placeholder={translated.text}
-        />
+        <Demo3 />
         <h2>{translated.title1}</h2>
-        <Input type="text" placeholder={translated.text} />
-        <Input type="password" placeholder={translated.password} />
-        <Input type="digit" placeholder={translated.digit} />
-        <Input type="number" placeholder={translated.number} />
+        <Demo4 />
         <h2>{translated.title2}</h2>
-        <Input readOnly placeholder={translated.readOnly} />
-        <Input disabled placeholder={translated.disabled} />
+        <Demo5 />
         <h2>{translated.title3}</h2>
-        <Input clearable placeholder={translated.clear} />
-        <Input clearable clearIcon={<Close />} placeholder={translated.clear} />
+        <Demo6 />
         <h2>{translated.withForm}</h2>
-        <Form>
-          <Form.Item label={translated.text1} name="username">
-            <Input
-              className="nut-input-text"
-              placeholder={translated.text}
-              type="text"
-              onChange={(val) => {
-                console.log('change value:', val)
-              }}
-            />
-          </Form.Item>
-        </Form>
+        <Demo7 />
         <h2>{translated.wordCount}</h2>
-        <div
-          style={{
-            display: 'flex',
-            alignItems: 'center',
-            background: '#fff',
-            padding: '0 10px',
-          }}
-        >
-          <Input
-            type="text"
-            maxLength={20}
-            onChange={(val) => setCurrentLength(val.length)}
-          />
-          <div className="right" style={{ fontSize: '12px' }}>
-            {currentLength} / 20
-          </div>
-        </div>
+        <Demo8 />
         <h2>{translated.password1}</h2>
-        <div
-          style={{
-            display: 'flex',
-            alignItems: 'center',
-            background: '#fff',
-            padding: '0 10px',
-          }}
-        >
-          <Input type={inputType} placeholder={translated.password} />
-          <div
-            className="right"
-            onClick={() =>
-              setInputType(inputType === 'text' ? 'password' : 'text')
-            }
-          >
-            {inputType === 'text' ? (
-              <Eye color="var(--nutui-gray-7)" />
-            ) : (
-              <Marshalling color="var(--nutui-gray-7)" />
-            )}
-          </div>
-        </div>
+        <Demo9 />
         <h2>{translated.title6}</h2>
-        <Input formatter={formatter} placeholder={translated.formatter} />
-        <Input
-          formatter={formatter}
-          formatTrigger="onBlur"
-          placeholder={translated.formatter2}
-        />
+        <Demo10 />
         <h2>{translated.title8}</h2>
-        <Input align="left" placeholder={translated.align} />
-        <Input align="right" placeholder={translated.align} />
+        <Demo11 />
         <h2>{translated.title10}</h2>
-        <Input
-          placeholder={translated.title10}
-          onClick={() => Toast.show('onClick')}
-        />
+        <Demo12 />
         <h2>{translated.title11}</h2>
-        <div
-          style={{
-            display: 'flex',
-            alignItems: 'center',
-            background: '#fff',
-            padding: '0 10px',
-          }}
-        >
-          <Tips color="var(--nutui-gray-7)" />
-          <Input
-            placeholder={translated.codeplaceholder}
-            style={{ '--nutui-input-padding': '10px' }}
-          />
-          <div className="right">
-            <Button type="primary" size="small">
-              {translated.sendCode}
-            </Button>
-          </div>
-        </div>
+        <Demo13 />
         <h2>{translated.border}</h2>
-        <Input
-          style={{ '--nutui-input-border-bottom-width': '1px' }}
-          placeholder={translated.border}
-        />
+        <Demo14 />
       </div>
     </>
   )
