+++ src/packages/input/demos/h5/demo11.tsx
@@ -1,0 +1,19 @@
+/*
+ * @Author: yeyu98
+ * @Date: 2024-03-15 12:45:23
+ * @LastEditors: yeyu98
+ * @LastEditTime: 2024-03-15 12:47:11
+ * @Description:
+ */
+import React from 'react'
+import { Input } from '@nutui/nutui-react'
+
+const Demo11 = () => {
+  return (
+    <>
+      <Input align="left" placeholder="文本内容对齐" />
+      <Input align="right" placeholder="文本内容对齐" />
+    </>
+  )
+}
+export default Demo11
