+++ src/packages/input/demos/h5/demo14.tsx
@@ -1,0 +1,21 @@
+/*
+ * @Author: yeyu98
+ * @Date: 2024-03-15 12:45:23
+ * @LastEditors: yeyu98
+ * @LastEditTime: 2024-03-15 12:50:08
+ * @Description:
+ */
+import React from 'react'
+import { Input } from '@nutui/nutui-react'
+
+const Demo14 = () => {
+  return (
+    <>
+      <Input
+        style={{ '--nutui-input-border-bottom-width': '1px' }}
+        placeholder="边框"
+      />
+    </>
+  )
+}
+export default Demo14
