+++ src/packages/input/demos/h5/demo3.tsx
@@ -1,3 +1,10 @@
+/*
+ * @Author: yeyu98
+ * @Date: 2024-03-15 12:26:16
+ * @LastEditors: yeyu98
+ * @LastEditTime: 2024-03-15 12:34:43
+ * @Description:
+ */
 import React, { useState } from 'react'
 import { Input } from '@nutui/nutui-react'
 
@@ -8,7 +15,7 @@ const Demo3 = () => {
       <Input
         value={val}
         onChange={(val) => setVal(val)}
-        placeholder={translated.text}
+        placeholder="请输入文本"
       />
     </>
   )
