+++ src/packages/input/demos/h5/demo7.tsx
@@ -1,16 +1,29 @@
+/*
+ * @Author: yeyu98
+ * @Date: 2024-03-15 12:26:16
+ * @LastEditors: yeyu98
+ * @LastEditTime: 2024-03-15 12:41:56
+ * @Description:
+ */
 import React from 'react'
 import { Input } from '@nutui/nutui-react'
+import Form from '@/packages/form'
 
 const Demo7 = () => {
-  const formatter = (value: string) => value.replace(/\d/g, '')
   return (
     <>
-      <Input formatter={formatter} placeholder="在输入时执行格式化" />
-      <Input
-        formatter={formatter}
-        formatTrigger="onBlur"
-        placeholder="在失焦时执行格式化"
-      />
+      <Form>
+        <Form.Item label="文本" name="username">
+          <Input
+            className="nut-input-text"
+            placeholder="请输入文本"
+            type="text"
+            onChange={(val) => {
+              console.log('change value:', val)
+            }}
+          />
+        </Form.Item>
+      </Form>
     </>
   )
 }
